Clean up EEPROM interfaces
