+++ Marlin/src/HAL/STM32_F4_F7/eeprom_wired.cpp
@@ -35,10 +35,10 @@
 #include "../shared/eeprom_api.h"
 
 size_t PersistentStore::capacity()    { return E2END + 1; }
-bool PersistentStore::access_start()  { return true; }
 bool PersistentStore::access_finish() { return true; }
 
 bool PersistentStore::access_start()  {
+  eeprom_init();
   return true;
 }
